Add weakness and clan storage
