+++ src/test/java/antafes/vampireEditor/TestClanUtility.java
@@ -27,11 +27,18 @@ import antafes.vampireEditor.entity.EntityStorageException;
 import antafes.vampireEditor.entity.character.Clan;
 import antafes.vampireEditor.entity.storage.AdvantageStorage;
 import antafes.vampireEditor.entity.storage.StorageFactory;
+import antafes.vampireEditor.entity.storage.WeaknessStorage;
+import org.testng.annotations.BeforeMethod;
 
 /**
  * Test clan utility class.
  */
 public class TestClanUtility {
+    @BeforeMethod
+    public void setUp() {
+        StorageFactory.storageWarmUp();
+    }
+
     /**
      * Create a test clan.
      *
@@ -40,12 +47,13 @@ public class TestClanUtility {
      */
     public static Clan createTestClan() throws EntityException, EntityStorageException {
         AdvantageStorage advantageStorage = (AdvantageStorage) StorageFactory.getStorage(StorageFactory.StorageType.ADVANTAGE);
+        WeaknessStorage weaknessStorage = (WeaknessStorage) StorageFactory.getStorage(StorageFactory.StorageType.WEAKNESS);
         return new Clan.Builder()
             .setKey("testClan")
             .addName(Configuration.Language.ENGLISH, "Test clan")
             .addNickname(Configuration.Language.ENGLISH, "Test nickname")
             .addAdvantage(advantageStorage.getEntity("allies"))
-            .addWeakness(VampireEditor.getWeakness("bloodTithe"))
+            .addWeakness(weaknessStorage.getEntity("bloodTithe"))
             .build();
     }
 }
