+++ src/test/java/antafes/vampireEditor/VampireEditorTest.java
@@ -23,9 +23,8 @@
 package antafes.vampireEditor;
 
 import antafes.vampireEditor.entity.EntityException;
-import antafes.vampireEditor.entity.EntityStorageException;
 import antafes.vampireEditor.entity.character.*;
-import antafes.vampireEditor.entity.storage.AdvantageStorage;
+import antafes.vampireEditor.entity.storage.ClanStorage;
 import antafes.vampireEditor.entity.storage.StorageFactory;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -77,43 +76,12 @@ public class VampireEditorTest {
 
     public void testGetClans() {
         new VampireEditor();
-        final HashMap actual = VampireEditor.getClans();
+        ClanStorage clanStorage = (ClanStorage) StorageFactory.getStorage(StorageFactory.StorageType.CLAN);
+        final HashMap actual = clanStorage.getList();
 
         Assert.assertNotNull(actual);
         Assert.assertFalse(actual.isEmpty());
         Assert.assertEquals(actual.values().toArray()[0].getClass(), Clan.class);
-    }
-
-    public void testGetWeakness() throws EntityException {
-        new VampireEditor();
-        final Weakness expected = new Weakness.Builder()
-            .setKey("bloodTithe")
-            .addName(Configuration.Language.ENGLISH, "Blood tithe 20%")
-            .addName(Configuration.Language.GERMAN, "Blutzehnt 20%")
-            .build();
-        final Weakness actual = VampireEditor.getWeakness("bloodTithe");
-
-        Assert.assertEquals(actual, expected);
-    }
-
-    public void testGetClan() throws EntityException, EntityStorageException {
-        new VampireEditor();
-        AdvantageStorage storage = (AdvantageStorage) StorageFactory.getStorage(StorageFactory.StorageType.ADVANTAGE);
-        final Clan expected = new Clan.Builder()
-            .setKey("assamites")
-            .addName(Configuration.Language.ENGLISH, "Assamites")
-            .addName(Configuration.Language.GERMAN, "Assamiten")
-            .addNickname(Configuration.Language.ENGLISH, "Saracens")
-            .addNickname(Configuration.Language.GERMAN, "Sarazenen")
-            .addAdvantage(storage.getEntity("auspex"))
-            .addAdvantage(storage.getEntity("presence"))
-            .addAdvantage(storage.getEntity("quietus"))
-            .addWeakness(VampireEditor.getWeakness("bloodTithe"))
-            .addWeakness(VampireEditor.getWeakness("diableryTraces"))
-            .build();
-        final Clan actual = VampireEditor.getClan("assamites");
-
-        Assert.assertEquals(actual, expected);
     }
 
     public void testGetMerits() {
